GAME: updated, too many loop added.
