+++ srcs/requirements/backend/nest-js/src/game/game.service.ts
@@ -17,114 +17,6 @@ export class GameService {
 		private readonly	usersService: UsersService,
 	) {}
 
-	restartBall(
-		grd: Game
-	)
-	{
-		let	katsayi: number = 1;
-		const newPos = Math.floor(Math.random() * 2);
-		const moves = [
-			{ x: -4, y: -2 },
-			{ x: -3, y: -4 },
-			{ x: -3, y: 3 },
-			{ x: -3, y: 2 },
-			{ x: 4, y: -2 },
-			{ x: 2, y: -4 },
-		];
-		if (grd.mode === 'fast-mode')
-			katsayi = 1.5;
-		const initialMove = moves[Math.floor(Math.random() * moves.length)];
-		grd.ball = { x: 479, speedX: initialMove.x * katsayi, speedY: initialMove.y * katsayi };
-		if (newPos == 0)
-			grd.ball.y = 22;
-		else
-			grd.ball.y = 757;
-	}
-
-	calcGameLoop(grd: Game): {
-		data: ILiveData;
-	}{
-		if (!grd)
-			return ;
-
-		// zaman bittiginde calisan yer
-		if (grd.duration <= 0)
-			grd.running = false;
-
-		// top hareketi
-		// if (grd.ball.y + 42 >= 800
-		if (grd.ball.y + 42 >= 800 || grd.ball.y <= 0)
-			grd.ball.speedY *= -1;
-	
-		if (grd.ball.y + 21 >= grd.playerL.location
-			&& grd.ball.y + 21 <= grd.playerL.location + 120)
-		{ //Left Player Collision
-			if (grd.ball.x <= 10)
-				{
-					grd.ball.speedX *= -1;
-					grd.ball.speedY += grd.playerL.speed / 10;
-					grd.ball.x += 3;
-				}
-		}
-		else if (grd.ball.y + 21 >= grd.playerR.location
-			&& grd.ball.y + 21 <= grd.playerR.location + 120)
-		{ //Right Player Collision
-			if (grd.ball.x + 42 >= 990)
-			{
-				grd.ball.speedX *= -1;
-				grd.ball.speedY += grd.playerR.speed / 10;
-				grd.ball.x -= 3;
-			}
-		}
-
-		//add score to left
-		if (grd.ball.x + 42 >= 1000)
-		{
-			grd.playerL.score += 1;
-			if (grd.playerL.score >= grd.winScore){
-				grd.running = false;
-			}
-			this.restartBall(grd);
-		}
-		//add score to right
-		else if (grd.ball.x <= 0)
-		{
-			grd.playerR.score += 1;
-			if (grd.playerR.score >= grd.winScore){
-				grd.running = false;
-			}
-			this.restartBall(grd);
-		}
-
-		grd.ball.x += grd.ball.speedX;
-		grd.ball.y += grd.ball.speedY;
-		
-		// sol oyuncu hareketi
-		if (grd.playerL.location + grd.playerL.speed <= 0)
-			grd.playerL.location = 0;
-		else if (grd.playerL.location + grd.playerL.speed + 120 >= 800)
-			grd.playerL.location = 800 - 120;
-		else
-			grd.playerL.location += grd.playerL.speed;
-
-		// sag oyuncu hareketi
-		if (grd.playerR.location + grd.playerR.speed <= 0)
-			grd.playerR.location = 0;
-		else if (grd.playerR.location + grd.playerR.speed + 120>= 800)
-			grd.playerR.location = 800 - 120;
-		else
-			grd.playerR.location += grd.playerR.speed;
-
-		const	returnData: ILiveData = {
-			ball: grd.ball,
-			playerL: grd.playerL,
-			playerR: grd.playerR,
-			duration: grd.duration,
-			running: grd.running,
-		};
-		return ({data: returnData});
-	}
-
 	async	transferPlayer(
 		{user}
 		: { user: User }
