+++ srcs/requirements/frontend/react-app/src/game/GameLobby.tsx
@@ -46,6 +46,7 @@ export interface ILobby {
 	ball: IBall,
 	playerLeft: IUser,
 	playerRight: IUser,
+	running: boolean,
 }
 
 const GameLobby = () => {
